+++ packages/build-utils/src/types.ts
@@ -447,6 +447,18 @@ export interface ProjectSettings {
   commandForIgnoringBuildStep?: string | null;
 }
 
+/*
+ * This is a builder whose build output version may dynamically change.
+ */
+export interface BuilderVX {
+  version: -1;
+  build: BuildVX;
+  diagnostics?: Diagnostics;
+  prepareCache?: PrepareCache;
+  shouldServe?: ShouldServe;
+  startDevServer?: StartDevServer;
+}
+
 export interface BuilderV2 {
   version: 2;
   build: BuildV2;
@@ -622,6 +634,10 @@ export interface BuildResultV2Typical {
   deploymentId?: string;
 }
 
+export type BuildResultVX =
+  | { resultVersion: 2; result: BuildResultV2 }
+  | { resultVersion: 3; result: BuildResultV3 };
+
 export type BuildResultV2 = BuildResultV2Typical | BuildResultBuildOutput;
 
 export interface BuildResultV3 {
@@ -630,6 +646,7 @@ export interface BuildResultV3 {
   output: Lambda | EdgeFunction;
 }
 
+export type BuildVX = (options: BuildOptions) => Promise<BuildResultVX>;
 export type BuildV2 = (options: BuildOptions) => Promise<BuildResultV2>;
 export type BuildV3 = (options: BuildOptions) => Promise<BuildResultV3>;
 export type PrepareCache = (options: PrepareCacheOptions) => Promise<Files>;