+++ packages/elysia/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/elysia
 
+## 0.1.52
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.19
+
 ## 0.1.51
 
 ### Patch Changes