+++ packages/node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/node
 
+## 5.6.19
+
+### Patch Changes
+
+- Updated dependencies [[`e5564b684ee40dd60f745327ef206bbe58a5191a`](https://github.com/vercel/vercel/commit/e5564b684ee40dd60f745327ef206bbe58a5191a)]:
+  - @vercel/build-utils@13.9.0
+
 ## 5.6.18
 
 ### Patch Changes